corrected updates as discussed

+++ eKhaya/9.1.0/aspnet-core/src/eKhaya.Core/Domain/Payments/Payment.cs
@@ -17,7 +17,7 @@ namespace eKhaya.Domain.Payment
 
         public virtual Resident Tenant { get; set; }
 
-        public virtual double Amount { get; set; }
+        public virtual decimal Amount { get; set; }
 
 
     }
