+++ eKhaya/9.1.0/aspnet-core/src/eKhaya.Core/Domain/Properties/Property.cs
@@ -15,7 +15,7 @@ namespace eKhaya.Domain.Properties
     {
         public virtual string Address { get; set; }
 
-        public virtual int Size { get; set; }
+        public virtual decimal Size { get; set; }
         
         public virtual ICollection<Image> Images { get; set; }
